+++ src/quick-lint-js/fe/parse-type.cpp
@@ -704,23 +704,14 @@ again:
   }
 
   // import("module").Name
-  case Token_Type::kw_import: {
-    Source_Code_Span import_keyword_span = this->peek().span();
+  case Token_Type::kw_import:
     this->skip();
     QLJS_PARSER_UNIMPLEMENTED_IF_NOT_TOKEN(Token_Type::left_paren);
     this->skip();
     QLJS_PARSER_UNIMPLEMENTED_IF_NOT_TOKEN(Token_Type::string);
     this->skip();
     QLJS_PARSER_UNIMPLEMENTED_IF_NOT_TOKEN(Token_Type::right_paren);
-    const Char8 *right_paren_end = this->peek().end;
     this->skip();
-    if (this->peek().type != Token_Type::dot) {
-      this->diag_reporter_->report(
-          Diag_TypeScript_Import_Type_Missing_Export_Name{
-              .expected_export_name = Source_Code_Span::unit(right_paren_end),
-              .import_keyword = import_keyword_span,
-          });
-    }
     while (this->peek().type == Token_Type::dot) {
       this->skip();
       switch (this->peek().type) {
@@ -740,7 +731,6 @@ again:
     }
     maybe_parse_dots_after_generic_arguments();
     break;
-  }
 
   case Token_Type::comma:
   case Token_Type::end_of_file: